realloc: follow up for the moving case
Lucian correctly points out that the prior fix to realloc,
db4644436d794719830fb97d5d96d4f9ae876f89, didn't go far enough.  While
it correctly handled the case of rederiving bounds for the non-moving
case, it failed to consider that, in the moving case, the old cap bounds
might be less than the sizeclass's size as well.  In this case, the
memcpy, which uses the passed user pointer, will attempt to reach out of
these bounds.

On CHERI, we can avoid the use of ->alloc_size() and just grab the
length.  This might, amusingly, let a user get an object of a smaller
sizeclass into this bin, but the local sz value is never pushed further
into the allocator: we re-enter the allocator through free() which will,
on CHERI, necessarily rederive a privileged pointer and the correct size
class information.