caprevoke policy tweak to improve concurrency
I never really meant to be synchronizing when sealing a chunk, and yet I
was.  Don't quite take the fastest out option (IGNORE_START, NO_WAIT_OK,
LAST_NO_EARLY), but if there's an epoch open, volunteer to close it (so
IGNORE_START, NO_WAIT_OK, ONLY_IF_OPEN, LAST_PASS).  I'm not sure this
is the right policy, but in practice, we don't leave epochs open, so
this will amount to the same thing as the fast out call.