[FOR-UPSTREAM] Avoid linking the snmalloc shim against the C++ stdlib
We are not using any C++ symbols and linking against libc should be sufficient.
To achieve this we can tell CMake to link with cc instead of c++ by
settisetting the LINKER_LANGUAGE property to C.
This is very useful for benchmarking on minimal CHERI images that might
not contain libc++ (I have since added libc++ by default but this change
is still useful).